+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/storage/channel/FluidStorageChannelType.java
@@ -10,12 +10,17 @@ import com.refinedmods.refinedstorage2.platform.api.resource.list.FuzzyResourceL
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.AbstractPlatformStorageChannelType;
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.FuzzyStorageChannelImpl;
 import com.refinedmods.refinedstorage2.platform.common.Platform;
+import com.refinedmods.refinedstorage2.platform.common.internal.grid.view.FluidGridResource;
+import com.refinedmods.refinedstorage2.platform.common.screen.widget.AbstractSideButtonWidget;
 import com.refinedmods.refinedstorage2.platform.common.util.PacketUtil;
 
 import java.util.Optional;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.resources.ResourceLocation;
+
+import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createTranslation;
 
 class FluidStorageChannelType extends AbstractPlatformStorageChannelType<FluidResource> {
     FluidStorageChannelType() {
@@ -25,7 +30,8 @@ class FluidStorageChannelType extends AbstractPlatformStorageChannelType<FluidRe
                 final ResourceList<FluidResource> list = new ResourceListImpl<>();
                 final FuzzyResourceList<FluidResource> fuzzyList = new FuzzyResourceListImpl<>(list);
                 return new FuzzyStorageChannelImpl<>(fuzzyList);
-            }
+            },
+            createTranslation("misc", "storage_channel_type.fluid")
         );
     }
 
@@ -42,6 +48,26 @@ class FluidStorageChannelType extends AbstractPlatformStorageChannelType<FluidRe
     @Override
     public Optional<GridResource> toGridResource(final ResourceAmount<?> resourceAmount) {
         return Platform.INSTANCE.getFluidGridResourceFactory().apply(resourceAmount);
+    }
+
+    @Override
+    public boolean isGridResourceBelonging(final GridResource gridResource) {
+        return gridResource instanceof FluidGridResource;
+    }
+
+    @Override
+    public ResourceLocation getTextureIdentifier() {
+        return AbstractSideButtonWidget.DEFAULT_TEXTURE;
+    }
+
+    @Override
+    public int getXTexture() {
+        return 16;
+    }
+
+    @Override
+    public int getYTexture() {
+        return 128;
     }
 
     @Override